+++ packages/sdk/CHANGELOG.md
@@ -3,6 +3,7 @@
 - YYYY-MM-DD **BREAKING?** -- description
 
 - 2021-04-20 -- Removes **Deprecated** `params`, `buildParams`
+- 2021-04-21 -- Updates encoding naming of `gasLimit` and `script` to `computeLimit` and `cadence`. Internal only.
 
 ### 0.0.45-alpha.20 -- 2021-04-21
 
@@ -10,10 +11,10 @@
 
 ### 0.0.45-alpha.19 -- 2021-04-16
 
-- 2021-04-15 -- Add `createSignableVoucher` for message payload verification by wallet, pass as `voucher` to `PreSignable` and `Signable`
+- 2021-04-15 -- Adds `createSignableVoucher` for message payload verification by wallet, pass as `voucher` to `PreSignable` and `Signable`
 - 2021-04-15 -- Exposes `config` from SDK.
 - 2021-04-15 -- Removes use of `TextDecoder` from `send-get-account`.
-- 2021-04-13 -- Add `makeVoucher` for message verification by wallet, pass as `voucher` to `PreSignable` and `Signable`
+- 2021-04-13 -- Adds `makeVoucher` for message verification by wallet, pass as `voucher` to `PreSignable` and `Signable`
 - 2021-04-09 -- Internal only - Added `prepAccount` to `interaction`. Update `authorizations`, `payer`, `proposer` builders.
 
 ### 0.0.45-alpha.18 -- 2021-04-09