+++ packages/sdk/CHANGELOG.md
@@ -1,6 +1,10 @@
 ### Unreleased
 
 - YYYY-MM-DD **BREAKING?** -- description
+
+## 0.0.45 - 2021-04-16
+
+- 2021-04-16 -- Full VSN Release `@onflow/sdk` 0.0.45-alpha.20 -> 0.0.45
 - 2021-04-23 -- Moves type check utils to `/utils`
 - 2021-04-23 -- Move `createVoucher` to `/utils` to resolve circular dependency in `interaction`
 - 2021-04-22 -- Adds `wallet-utils` `validateSignableTransaction` support for wallets to validate Signable payload